refactor: extract frontend uuid helper
